+++ CHANGELOG.md
@@ -5,6 +5,11 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 
 ### [Unreleased]
 
+## [17.102.0-M4] - 2025-02-27
+### Changed
+- Split out the stream_error table into `stream_error` and `stream_error_hash`
+- StreamError Object now split into StreamErrorDetails and StreamErrorHash
+
 ## [17.102.0-M3] - 2025-02-18
 ### Changed
 - Errors set into stream_status table are now `upserts` to handle the case where no stream yet exists in stream_status